Update README.md
The project, named above, aims to train a model to drive a car independently on a track created. The power of a driving model is learned by combining the behavior of a human driver. Training data is obtained from human driving models in this simulator, and is then incorporated into a deep learning network that reads feedback (directional angle) throughout the frame encountered in the simulation. In other words, the model is trained to predict the appropriate directional angle for each frame while driving. The model is then verified in a new track to check the performance of the included components to make the directional angle prediction.